requirements.txt
flask
gunicorn
requests
anthropic
Pillow
